services/jataware_auto_legend/external/auto_legend/pyproject.toml
[tool.poetry]
name = "auto_legend"
version = "0.0.1"
description = ""
authors = ["bkj <ben@jataware.com>", "zeek <zeek@jataware.com>"]
readme = "README.md"
packages = [
        {include = "auto_legend", from = "."},
 ]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "2.1.3"
torch = "2.4.1"
pillow = "^11.0.0"
anthropic = "^0.39.0"
albumentations = "1.4.21"
torchvision = "0.19.1"
transformers = "4.45.2"
timm = "^1.0.11"

[tool.poetry.group.dev.dependencies]
rich = "^13.9.4"
rich-pixels = "^3.0.1"
pandas = "^2.2.3"
httpx = "^0.27.2"
scikit-learn = "^1.5.2"

[tool.poetry.scripts]
auto_legend = "auto_legend.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"